Not used AvroSchemaDeterminer removal
